+++ ssh2/__init__.py
@@ -1,3 +1,10 @@
 from .client import SSH
-from .core import ssh_connect
-from .errors import SSHConfigurationError, SSHConnectionError, SSHChannelError, SFTPError
+from .config import SSHConfigData
+from .core import SSHConnect
+from .core import SSHContext
+from .core import SSHTunnelContext
+from .errors import SSHConfigurationError
+from .errors import SSHConnectionError
+from .errors import SSHChannelError
+from .errors import SFTPError
+from .errors import SSHContextError