+++ ssh2/errors.py
@@ -1,6 +1,7 @@
 class SSHConfigurationError(RuntimeError):
     """
-    Exception raised when one or more unsupported SSH configuration properties are invoked.
+    Exception raised when one or more unsupported SSH configuration 
+    properties are invoked.
     """
     pass
 
@@ -14,15 +15,22 @@ class SSHConnectionError(RuntimeError):
 
 class SSHChannelError(SSHConnectionError):
     """
-    Exception raised when ``paramiko.open_channel`` fails to create a socket object for our
-    tunnel.
+    Exception raised when ``paramiko.open_channel`` fails to create a 
+    socket object for our tunnel.
     """
     pass
 
 
 class SFTPError(SSHConnectionError):
     """
-    Exception raised when an ``SSHClient`` object doesn't exist and the user attempts to create
-    a new ``SFTPClient`` session object.
+    Exception raised when an ``SSHClient`` object doesn't exist and the 
+    user attempts to create a new ``SFTPClient`` session object.
+    """
+    pass
+
+
+class SSHContextError(SSHConnectionError):
+    """
+    Exception raised when the SSH context cannot be created.
     """
     pass